tried adding bot and extra play button
i tried to add a bot paddle but it just floats away. I tried to make it not get further than the wall but i wasn't able to get it. Same for the play button, i have one. But i couldn't get it to send u to the 1 player mode.